move terraDecimalValidator
